fix: fix signature

+++ generate.js
@@ -52,7 +52,7 @@ const dataList = [
 ];
 
 const leafNodes = dataList.map(node => {
-    return ethers.utils.solidityKeccak256(['address', 'uint256', 'uint256'], [node.address, node.amount, node.index])
+    return ethers.utils.solidityKeccak256(['address', 'uint256', 'uint8'], [node.address, node.amount, node.index])
 });
 
 const merkletree = new MerkleTree(leafNodes, keccak256, { sortPairs: true });
@@ -66,7 +66,7 @@ module.exports = (address, index) => {
         return node.address.toLowerCase() === (address || '').toLowerCase() && node.index === +index
     });
     if (!node) return { proof: [], amount: 0 };
-    const leaf = ethers.utils.solidityKeccak256(['address', 'uint256', 'uint256'], [node.address, node.amount, node.index]);
+    const leaf = ethers.utils.solidityKeccak256(['address', 'uint256', 'uint8'], [node.address, node.amount, node.index]);
 
     return { proof: merkletree.getHexProof(leaf), amount: node.amount };
 }
